Code cleanup for Day 5
Greatly simplified the parsing of the file. Broke the file up into
blocks, rather than the way I looked for blank lines in my loop.
Also took more advantage of split on the strings. And learned about
the map function (a bit nicer than list comprehension for this case).
Because of switching to reading blocks, I put the example input in
a file to properly parse the input into blocks.

Switched from maps to parallel lists for the seeds and resulting
values. It just became easier, rather than trying to initialize the
map values to sys.maxsize. This also made it easier to use the min()
function with part 2.

Didn't like "inside", because it wasn't necessarily inside anything,
but overlap from the two ranges. So, switched it to "overlap".

Instead of using the size and doing math with the Mapping and
SeedRange, I just did it in the class and used a proprty for the end.

Added a method on Mapping to map a value to a destination, just to
have the logic in one place.

Took inspiration for file parsing from HyperNeutrino:
https://youtu.be/NmxHw_bHhGM?si=tQ2Pp8ugQqk00xXt