Code cleanup for Day 5
Cleaned up the logic for the range checking. Walked through the
execution of the original code to better understand it, since it
wasn't mine to begin with and confused me a lot. The new logic
is much simpler, I think.

We really only care if there is overlap. If there is overlap,
add it to the "answer" and only then do we check to see if there
is anything before or after it. If there is some of the range
before the overlap, then it should start with the original start
and go to the start of the overlap. If there is some of the range
after the overlap, then it should start with the end of the
overlap and go to the end of the original range.

If there is no overlap, just put the original range on the "new
range" list. The original logic relied on that range being part of
either the before or after and that was confusing. Now, you know
for sure that it is just dealing with the original range.