Solution for Day 8
Another not too difficult puzzle. Started with brute force for
part 2, but realized it didn't work. Couldn't remember the least
common multiple algorithm at first. Thankful that I found it and
the fact that python has a built-in function.

After watching HyperNeutrino's video for today, I realized I had
an incorrect assumption in part 2 that didn't actually hurt me.
The assumption was that if I had to keep going from the end, the
next iteration of the path was the same distance going Z to Z as
going from A to Z. This is what the example show, but I didn't
properly think that through. In theory, if it takes 10 steps to
get from A to Z, it could take 3 to get to the next Z. If I
was being more thorough, I should have figured out steps to the
next Z and to see if there is possibly any loop back to one of
the Z nodes, if not the first Z node. Because the puzzle did
have the same length, my lcm function worked. If that wasn't the
case, there would have needed to be a lot more computing of moves
and computations to get there. I think I could come up with a
better way of doing it than brute force, but I'm glad the
original idea acutally worked. The video is here:
https://youtu.be/_nnxLcrwO_U?si=QqUxP1yRvThlxshG

Started early this morning with a break in between.
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
  8   08:08:30  36989      0   09:46:33  27782      0