Solution for Day 10 Part 1
Part 1 took some time, but I actually got it. Part 2 has been hell
trying to figure out. I'm guessing that I'm doing things wrong and
it would be a lot easier if I did it a different way.

I did find out that I need to use "point in polygon", but need to
figure out the type of pipe the start is.

So, I'm committing part 1 for now and will continue to work on part
2.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 10   01:54:27   7524      0          -      -      -