Code cleanup for Day 11
Came up with a MUCH better way to process everything. Rather than
doing everything by brute force, keep track of where the galaxies
are and when we expand the universe, modify the coordinates of
the galaxy. One key thing is to look at the universe expansion in
reverse order, because you could have a galaxy move one too many
times. Ultimately, I'm able to use the same function for both parts,
with passing in the size of the expansion.