Solution for Day 14
Part 2 was a little more difficult. Obviously brute force of 1
billion would not be right. Thanks to Scott for cluing me in that
there was a loop of positions. Took a little bit to make a generic
tilt capability. Then it took a bit to get my math right for the
delta that I needed to do more cycles.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 14   07:23:48  18822      0   13:47:41  17979      0