Solution for Day 16
For a weekend, not too difficult. Two problems that I had as I was
working on it:
* not having the right directions on the angles (I think I screwed
  every single one originally)
* getting into a loop

Added the seen map to prevent the endless loop.

Once I got part 1 done, part 2 was easy.

Didn't finish it until Sunday because of family activities.
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 16       >24h  26198      0       >24h  25350      0