Solution for Day 17
This is what I was expecting for a weekend puzzle. Wasn't really
sure how I handled the limitation of a maximum of 3 blocks. Saw
something on Reddit about maintaining the distance and that helped.
Part 2 actually was realatively easy, with knowing how to do part
1.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 17   18:11:32  13972      0   18:35:46  12940      0