Solution for Day 18
A bit of a tough one. Was able to find out about the shoelace
formula for finding the area of a polygon, but it took reddit and
HyperNeutrino on YouTube to know about using Pick's Theorem. Didn't
think about the fact that the coordinates are from the middle of the
squares and that we miss area on the outside.

HyperNeutrino YouTube video:
https://youtu.be/bGWK76_e-LM?si=U-drVUytFow8Hxwg
Reddit post:
https://www.reddit.com/r/adventofcode/comments/18l8mao/2023_day_18_intuition_for_why_spoiler_alone/

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 18   13:17:51  16516      0   13:22:22  11159      0