Solution for Day 19
Part 1 was relatively easy with just traversing the rules. For
part 2, I made an assumption based on the example input that
there wouldn't be multiple conditions in a rule for an accept.
So, I kept finding the first accept condition and not the others.
I added numbers to the accept results and that got me past it.
I'm sure there's a different way to handle it, but this works.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 19   01:07:32   4363      0   16:25:04  12327      0