Solution for Day 20
Kind of a tough one. Took a little to get part 1 right. For part
2, I realized trying to go through it wouldn't work. First thought
about some caching of states, but that would mean there was some
loop and that probably wouldn't work. Started drawing the graph
from the bottom and it came to me that if I could figure out how
many button pushes it takes for each node that goes in "gf" to be
high, I could use least common multiple. Once I figure that out,
it actually was pretty easy. My solution is hard coded to my input
data for now. I was thinking about a way to make it more generic.
Probably in my code cleanup.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 20   13:18:13  12069      0   16:14:30   9470      0