Solution for Day 24 Part 1
Didn't start working on this until Christmas Day and struggled to
understand it, at first. Finally, focused on it today and figured
out what it was asking for. Ultimately, it was basic algebra,
which I figured out, on my own. Did find a page that simplified
the line_cross function for me:
http://gamedev.docrobs.co.uk/collisions-between-objects-using-line-intersection

Part 2 seems a little harder and will continue to work on it later.
Just wanted to get this committed now.

      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 24       >24h  13919      0          -      -      -